+++ include/nstd/os/windows/io.h
@@ -40,12 +40,14 @@ NSTDAPI NSTDOSWindowsIOHandle nstd_os_windows_io_stdout();
 
 /// Writes a buffer of `bytes` to `stream`.
 /// Parameters:
-///     `const NSTDOSWindowsIOHandle stream` - Handle to a standard IO stream.
+///     `const NSTDOSWindowsHandle stream` - Handle to a standard IO stream.
 ///     `const NSTDSlice *const bytes` - The bytes to write to the stream.
+///     `NSTDUInt32 *const written` - Returns as the number of bytes actually written.
 /// Returns: `NSTDErrorCode errc` - Nonzero on error.
 NSTDAPI NSTDErrorCode nstd_os_windows_io_write(
-    const NSTDOSWindowsIOHandle stream,
-    const NSTDSlice *const bytes);
+    const NSTDOSWindowsHandle stream,
+    const NSTDSlice *const bytes,
+    NSTDUInt32 *const written);
 
 #ifdef NSTDCPP
 }