+++ src/os/windows/io.rs
@@ -42,7 +42,7 @@ pub unsafe extern "C" fn nstd_os_windows_io_handle_as_handle(
 pub unsafe extern "C" fn nstd_os_windows_io_print(cstr: *const NSTDChar) -> NSTDErrorCode {
     let stdout = nstd_os_windows_io_stdout();
     let bytes = crate::core::cstr::nstd_core_cstr_as_slice(cstr);
-    nstd_os_windows_io_write(stdout, &bytes)
+    nstd_os_windows_io_write(GetStdHandle(stdout), &bytes, std::ptr::null_mut())
 }
 
 /// Writes a C string to stdout with a newline character.
@@ -68,20 +68,22 @@ pub unsafe extern "C" fn nstd_os_windows_io_stdout() -> NSTDOSWindowsIOHandle {
 
 /// Writes a buffer of `bytes` to `stream`.
 /// Parameters:
-///     `const NSTDOSWindowsIOHandle stream` - Handle to a standard IO stream.
+///     `const NSTDOSWindowsHandle stream` - Handle to a standard IO stream.
 ///     `const NSTDSlice *const bytes` - The bytes to write to the stream.
+///     `NSTDUInt32 *const written` - Returns as the number of bytes actually written.
 /// Returns: `NSTDErrorCode errc` - Nonzero on error.
 #[inline]
 #[cfg_attr(feature = "clib", no_mangle)]
 pub unsafe extern "C" fn nstd_os_windows_io_write(
-    stream: NSTDOSWindowsIOHandle,
+    stream: NSTDOSWindowsHandle,
     bytes: &NSTDSlice,
+    written: *mut u32,
 ) -> NSTDErrorCode {
     (WriteConsoleA(
-        GetStdHandle(stream),
+        stream,
         bytes.ptr.raw,
         bytes.size as u32,
-        std::ptr::null_mut(),
+        written,
         std::ptr::null_mut(),
     ) == 0) as NSTDErrorCode
 }